인트 더블로 바꾸었음
대충 설명

+++ Assets/Test3.cs
@@ -7,7 +7,7 @@ public class Test3 : MonoBehaviour
     // Start is called before the first frame update
     void Start()
     {
-        int test4;
+        double test4;
         test4 = 1001;
     }
